restructuring and updating copyrights.
