+++ src/main/java/org/wso2/qa/testlink/extension/model/CarbonComponent.java
@@ -1,13 +1,13 @@
-/**
+/*
  * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- * <p>
+ *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -31,11 +31,11 @@ public class CarbonComponent {
         this.componentVersion = componentVersion;
     }
 
-    String getComponentName() {
+    public String getComponentName() {
         return componentName;
     }
 
-    String getComponentVersion() {
+    public String getComponentVersion() {
         return componentVersion;
     }
