+++ src/main/java/org/wso2/qa/testlink/extension/web/ContextListener.java
@@ -1,13 +1,13 @@
-/**
+/*
  * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- * <p>
+ *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -21,11 +21,11 @@ import org.apache.log4j.Logger;
 import org.apache.log4j.MDC;
 import org.wso2.qa.testlink.extension.model.Configurations;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Properties;
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
+import java.io.InputStream;
+import java.io.IOException;
+import java.util.Properties;
 
 /**
  * Servlet context listener which does the initializations tasks.
