+++ src/main/java/org/wso2/qa/testlink/extension/web/TestResultManagementServlet.java
@@ -1,13 +1,13 @@
-/**
+/*
  * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- * <p>
+ *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License.
  * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -18,21 +18,22 @@
 package org.wso2.qa.testlink.extension.web;
 
 import org.apache.commons.lang.StringUtils;
-import org.apache.log4j.Logger;
-import org.apache.log4j.MDC;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.slf4j.MDC;
+import org.wso2.qa.testlink.extension.core.TestResultsUpdater;
+import org.wso2.qa.testlink.extension.exception.TestResultsManagerException;
 import org.wso2.qa.testlink.extension.model.CarbonComponent;
-import org.wso2.qa.testlink.extension.model.TestResultsManagerException;
-import org.wso2.qa.testlink.extension.model.TestResultsUpdater;
 
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * Triggers TestLink update process
@@ -42,7 +43,7 @@ public class TestResultManagementServlet extends HttpServlet {
 
     private static final long serialVersionUID = 42L;
 
-    private static final Logger LOGGER = Logger.getLogger(TestResultManagementServlet.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(TestResultManagementServlet.class);
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
